clean up encrypt, move encrypt/decrypt into functions
