+++ ipa-core/src/bin/test_decrypt.rs
@@ -15,37 +15,13 @@ use ipa_core::{
     },
     hpke::{KeyRegistry, PrivateKeyOnly},
     report::{EncryptedOprfReport, EventType, OprfReport},
-    secret_sharing::{replicated::ReplicatedSecretSharing, SharedValue},
     test_fixture::Reconstruct,
 };
-
-pub trait Serializer {
-    /// Converts self into a CSV-encoded byte string
-    /// ## Errors
-    /// If this conversion fails due to insufficient capacity in `buf` or other reasons.
-    fn to_csv<W: Write>(&self, buf: &mut W) -> std::io::Result<()>;
-}
-
-#[cfg(any(test, feature = "test-fixture"))]
-impl<BK: SharedValue, TV: SharedValue, TS: SharedValue> Serializer for OprfReport<BK, TV, TS> {
-    fn to_csv<W: Write>(&self, buf: &mut W) -> std::io::Result<()> {
-        let is_trigger_report = self.event_type == EventType::Trigger;
-
-        // fmt::write is cool because it does not allocate when serializing integers
-        write!(buf, "{:?},", self.timestamp.left())?;
-        write!(buf, "{:?},", self.match_key.left())?;
-        write!(buf, "{},", u8::from(is_trigger_report))?;
-        write!(buf, "{:?},", self.breakdown_key.left())?;
-        write!(buf, "{:?}", self.trigger_value.left())?;
-
-        Ok(())
-    }
-}
 
 #[derive(Debug, Parser)]
 #[clap(name = "test_decrypt", about = "Test Decrypt")]
 #[command(about)]
-struct Args {
+struct DecryptArgs {
     /// Path to helper1 file to decrypt
     #[arg(long)]
     input_file1: PathBuf,
@@ -85,7 +61,11 @@ async fn build_hpke_registry(
 
 #[tokio::main]
 async fn main() -> Result<(), BoxError> {
-    let args = Args::parse();
+    let args = DecryptArgs::parse();
+    decrypt_and_reconstruct(args).await
+}
+
+async fn decrypt_and_reconstruct(args: DecryptArgs) -> Result<(), BoxError> {
     let file1 = File::open(args.input_file1)?;
     let file2 = File::open(args.input_file2)?;
     let file3 = File::open(args.input_file3)?;
