+++ ipa-core/src/hpke/mod.rs
@@ -17,7 +17,8 @@ mod registry;
 
 pub use info::Info;
 pub use registry::{
-    KeyPair, KeyRegistry, PrivateKeyOnly, PrivateKeyRegistry, PublicKeyOnly, PublicKeyRegistry,
+    KeyPair, KeyRegistries, KeyRegistry, PrivateKeyOnly, PrivateKeyRegistry, PublicKeyOnly,
+    PublicKeyRegistry,
 };
 
 use crate::{
